+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-masonry/watchers-masonry.module.ts
@@ -13,6 +13,7 @@ import {MatIconModule} from "@angular/material/icon";
 import {MatButtonModule} from "@angular/material/button";
 import {MatCheckboxModule} from "@angular/material/checkbox";
 import {WatchHistoryComponent} from './watchers/watch-history/watch-history.component';
+import {WatchDiffComponent} from "./watchers/watch-diff/watch-diff.component";
 
 
 @NgModule({
@@ -23,7 +24,8 @@ import {WatchHistoryComponent} from './watchers/watch-history/watch-history.comp
     JsonWatcherComponent,
     WatcherCardComponent,
     WatchHistoryComponent,
-    HistoryTemplateDirective
+    HistoryTemplateDirective,
+    WatchDiffComponent
   ],
   exports: [
     WatchersMasonryComponent,